Add basic Debian support.
Move build files to debian/ directory, use template control file to provide
support for different versions of Scala, and provide a top-level Makefile to
provide a normal build & install process.

Getting the build working requires a few other Debian-specific fixes. Most
importantly, we need to disable the Zinc compiler and jump through hoops to get
Maven not to write the user's home directory. This is necessary because building
with pbuilder in an isolated chroot environment disables write access to almost
everything outside the source directory.