Rework Debian build to build all Scala packages in a single build.
On Debian this is required to get all the output file, including the changes and
dsc files that include signatures from all the packages. In order to keep the
Scala versions that are built flexible, this constructs the debian/control file
dynamically and the Makefile rules iterate over all the versions extracted from
the control file.